# api/openapi.yml
openapi: 3.1.0
info:
  title: Advert Market API
  description: Telegram Mini App marketplace for channel advertising
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/v1/channels/{id}:
    get:
      tags:
      - Channels
      summary: Get channel detail
      description: Full channel info including pricing rules
      operationId: getDetail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Channel detail
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelDetailResponse"
        "404":
          description: Channel not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelDetailResponse"
    put:
      tags:
      - Channels
      summary: Update channel
      description: Update channel details (owner only)
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelUpdateRequest"
        required: true
      responses:
        "200":
          description: Channel updated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelResponse"
        "403":
          description: Not the owner of the channel
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelResponse"
        "404":
          description: Channel not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Channels
      summary: Deactivate channel
      description: Soft-delete a channel (owner only)
      operationId: deactivate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Channel deactivated
        "403":
          description: Not the owner of the channel
        "404":
          description: Channel not found
      security:
      - bearerAuth: []
...
  /api/v1/channels:
    get:
      tags:
      - Channels
      summary: Search channels
      description: Public channel catalog with filters and BM25 search
      operationId: search
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: minSubscribers
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minSubs
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: maxSubscribers
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: maxSubs
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: maxPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: minEngagement
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: language
        in: query
        required: false
        schema:
          type: string
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: SUBSCRIBERS_DESC
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: Search results
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CursorPageChannelListItem"
    post:
      tags:
      - Channels
      summary: Register a channel
      description: Registers a previously verified channel
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelRegistrationRequest"
        required: true
      responses:
        "201":
          description: Channel registered
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelResponse"
        "403":
          description: User is not an admin of the channel
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelResponse"
        "409":
          description: Channel already registered
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelResponse"
      security:
      - bearerAuth: []
...
  /api/v1/channels/verify:
    post:
      tags:
      - Channels
      summary: Verify channel for registration
      description: Checks that the bot is an admin and the user owns the channel
      operationId: verify
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelVerifyRequest"
        required: true
      responses:
        "200":
          description: Verification result
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelVerifyResponse"
        "403":
          description: Bot or user lacks required permissions
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelVerifyResponse"
        "404":
          description: Channel not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChannelVerifyResponse"
      security:
      - bearerAuth: []
...
  /api/v1/channels/count:
    get:
      tags:
      - Channels
      summary: Count channels
      description: Returns channel count for given search filters
      operationId: count
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: minSubscribers
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minSubs
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: maxSubscribers
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: maxSubs
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: maxPrice
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: minEngagement
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: language
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Channel count
          content:
            '*/*':
              schema:
                type: integer
                format: int64
...
components:
  schemas:
...
    ChannelUpdateRequest:
      type: object
      description: Channel update request
      properties:
        description:
          type: string
        categories:
          type: array
          items:
            type: string
        pricePerPostNano:
          type: integer
          format: int64
        language:
          type: string
          maxLength: 10
          minLength: 0
        isActive:
          type: boolean
    ChannelResponse:
      type: object
      description: Registered channel
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        username:
          type: string
        description:
          type: string
        subscriberCount:
          type: integer
          format: int32
        categories:
          type: array
          items:
            type: string
        pricePerPostNano:
          type: integer
          format: int64
        isActive:
          type: boolean
        ownerId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
...
    PricingRuleDto:
      type: object
      description: Channel pricing rule
      properties:
        id:
          type: integer
          format: int64
        channelId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        postTypes:
          type: array
          items:
            type: string
            enum:
            - REPOST
            - NATIVE
            - STORY
            - INTEGRATION
            - REVIEW
            - MENTION
            - GIVEAWAY
            - PINNED
            - POLL
            - FORWARD
          uniqueItems: true
        priceNano:
          type: integer
          format: int64
        isActive:
          type: boolean
        sortOrder:
          type: integer
          format: int32
...
    ChannelRegistrationRequest:
      type: object
      description: Channel registration request
      properties:
        channelId:
          type: integer
          format: int64
          description: Telegram chat ID from verify step
        categories:
          type: array
          description: Category slugs
          example:
          - tech
          items:
            type: string
        pricePerPostNano:
          type: integer
          format: int64
          description: Price per post in nanoTON
...
    ChannelVerifyRequest:
      type: object
      description: Channel verification request
      properties:
        channelUsername:
          type: string
          description: Public channel username without @
          example: mychannel
          maxLength: 255
          minLength: 0
      required:
      - channelUsername
    BotStatus:
      type: object
      description: Bot status in the channel
      properties:
        isAdmin:
          type: boolean
        canPostMessages:
          type: boolean
        canEditMessages:
          type: boolean
        missingPermissions:
          type: array
          items:
            type: string
    ChannelVerifyResponse:
      type: object
      description: Channel verification result
      properties:
        channelId:
          type: integer
          format: int64
        title:
          type: string
        username:
          type: string
        subscriberCount:
          type: integer
          format: int32
        botStatus:
          $ref: "#/components/schemas/BotStatus"
        userStatus:
          $ref: "#/components/schemas/UserStatus"
    UserStatus:
      type: object
      description: User status in the channel
      properties:
        isMember:
          type: boolean
        role:
          type: string
...
    ChannelListItem:
      type: object
      description: Channel search result item
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        username:
          type: string
        categories:
          type: array
          items:
            type: string
        subscriberCount:
          type: integer
          format: int32
        avgViews:
          type: integer
          format: int32
        engagementRate:
          type: number
        pricePerPostNano:
          type: integer
          format: int64
        isActive:
          type: boolean
        updatedAt:
          type: string
          format: date-time
    CursorPageChannelListItem:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChannelListItem"
        nextCursor:
          type: string
    ChannelDetailResponse:
      type: object
      description: Full channel detail with pricing rules
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        username:
          type: string
        description:
          type: string
        subscriberCount:
          type: integer
          format: int32
        categories:
          type: array
          items:
            type: string
        pricePerPostNano:
          type: integer
          format: int64
        isActive:
          type: boolean
        ownerId:
          type: integer
          format: int64
        engagementRate:
          type: number
        avgViews:
          type: integer
          format: int32
        language:
          type: string
        pricingRules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRuleDto"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT